feat: connect to Firesotre
read database on Snapshot
